django/requirements.txt
django
djangorestframework
mysqlclient
tzdata
